Refactor: Move Builder.Builder.InnerAsSpan to a separate file
